requirements_local.txt
# Dépendances pour l'installation locale
# Système de Gestion Immobilière

# Framework Web Principal
Flask==3.1.1
Werkzeug==3.0.4

# Base de données
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.35

# Formulaires et validation
Flask-WTF==1.2.2
WTForms==3.1.2
email-validator==2.2.0

# Génération PDF
reportlab==4.4.1
PyPDF2==3.0.1

# Utilitaires
python-dateutil==2.9.0
python-dotenv==1.0.1

# Services optionnels (SMS/Email)
sendgrid==6.11.0
twilio==9.2.4

# Serveur WSGI (pour production locale)
gunicorn==23.0.0

# Traductions (si utilisées)
translations==0.1.0